requirements.txt
streamlit==1.41.1
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
requests==2.31.0
gspread==6.1.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1
